* Bugfixes - Added missing dependency - Added logging - int must be Integer - do not guess string values (+ removed tests)

+++ src/main/java/ch/sama/sql/dialect/tsql/type/TYPE.java
@@ -190,7 +190,7 @@ public class TYPE {
             case "float":
                 return Double.class;
             case "int":
-                return Long.class;
+                return Integer.class;
             case "datetime":
                 return Date.class;
             case "uniqueidentifier":
